+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Testing: Upgrade nim-waku node to v0.3.
 - **Breaking**: Modify `WakuStore.queryHistory()` to accept one `Object` instead of multiple individual arguments.
 - `getStatusFleetNodes` return prod nodes by default, instead of test nodes.
+- Examples (web chat): Connect to prod fleet by default, test fleet for local development.
 
 ### Fixed
 - Expose `Enviroment` and `Protocol` enums to pass to `getStatusFleetNodes`.